Extract existing implementation into interface and abstract base class with lifecycle handling.
