+++ processors/db-processor/src/main/kotlin/net/corda/processors/db/internal/reconcile/db/ConfigReconciler.kt
@@ -35,7 +35,7 @@ class ConfigReconciler(
 
         if (dbReconciler == null) {
             dbReconciler =
-                DbReconcilerReader(
+                ClusterDbReconcilerReader(
                     coordinatorFactory,
                     dbConnectionManager,
                     String::class.java,
