+++ processors/db-processor/src/main/kotlin/net/corda/processors/db/internal/reconcile/db/CpiReconciler.kt
@@ -36,7 +36,7 @@ class CpiReconciler(
 
         if (dbReconciler == null) {
             dbReconciler =
-                DbReconcilerReader(
+                ClusterDbReconcilerReader(
                     coordinatorFactory,
                     dbConnectionManager,
                     CpiIdentifier::class.java,
