+++ processors/db-processor/src/main/kotlin/net/corda/processors/db/internal/reconcile/db/VirtualNodeReconciler.kt
@@ -36,7 +36,7 @@ class VirtualNodeReconciler(
 
         if (dbReconciler == null) {
             dbReconciler =
-                DbReconcilerReader(
+                ClusterDbReconcilerReader(
                     coordinatorFactory,
                     dbConnectionManager,
                     HoldingIdentity::class.java,
